Misc

Utils.unpickle — Function

unpickle(filename)

Read a Python data file using Python's Pickle via PyCall.jl.

Utils.subscript — Function

subscript(i::Integer)

Set up string with integer i as subscript (for printing).

From https://stackoverflow.com/questions/46671965/printing-variable-subscripts-in-julia

Examples

julia> println("Studio"*subscript(45))
Studio₄₅

Utils.pad_digits — Function

pad_digits(x)

Returns a vector with string representations of the numbers in x such that they have the same number of digits, including trailing zeros.

Useful when printing numbers in latex tables and annotating figures.

Examples

julia> pad_digits([1.21, 13.3, 123.456])
3-element Vector{String}:
 "1.210"
 "13.300"
 "123.456"

Utils.CovMatEquiCorr — Function

CovMatEquiCorr(σₓ, ρ, pBlock)

Set up a covariance matrix with equi-correlation within blocks of variables

σₓ is a p-vector with standard deviations

ρ[i] is the correlation within block i

pBlock[j] is the number of variables in block j

Examples

julia> σₓ = [1,2,3]; ρ = [0.5,0.8]; pBlock = [1,2];
julia> CovMatEquiCorr(σₓ, ρ, pBlock)
3×3 Matrix{Float64}:
 1.0  0.0  0.0
 0.0  4.0  4.8
 0.0  4.8  9.0

Utils.Cov2Corr — Function

ρ, σ = Cov2Corr(Σ)

Compute the correlation matrix ρ and vector standard deviations σ for the covariance matrix Σ.

Examples

julia> Σ = CovMatEquiCorr([1,2,3], [0.7], [3]) # Covariance matrix with all corr = 0.7
3×3 Matrix{Float64}:
 1.0  1.4  2.1
 1.4  4.0  4.2
 2.1  4.2  9.0

julia> ρ, σ = Cov2Corr(Σ); ρ
3×3 Matrix{Float64}:
 1.0  0.7  0.7
 0.7  1.0  0.7
 0.7  0.7  1.0

Utils.plotFcnGrid — Function

plotFcnGrid(f, xGrid, xNames, fcnArgs...;ylabel="", title ="", levels = 10,fill=:viridis)

Plotting a function of a 1D or 2D grid.

Utils.plotClassifier2D — Function

plotClassifier2D(y, X, predictFunc; gridSize = [100,100], colors = missing, axisLabels = missing)

Plots the decision boundaries of a classifier predictFunc() with two inputs

predictFunc should take x1, x2 as inputs and returns a class in unique(y)

class labels can be Categorical or Int